refactor(app): centralize patchset details state construction
This commit moves patchset details state assembly into
PatchsetDetailsState::from_rendered_preview. App::open_patchset_details now
focuses on orchestrating Lore and Render actor calls while the details screen
state owns conversion from rendered preview strings into the existing Ratatui
text payloads.

This commit is part of the architecture's refactoring phase 7.
